refactor: move AsyncId to deno_core and remove stale comment
Address review feedback from @bartlomieju:
- Move AsyncId type from ext/node/ops/handle_wrap.rs to
  libs/core/uv_compat.rs since it's a libuv concept
- Initialize AsyncId in JsRuntime::new_inner alongside the uv loop
- Remove stale comment about uv loop auto-creation in ext/node/lib.rs
- Re-export AsyncId from handle_wrap.rs for backwards compatibility
